update the metadata setting and update og image url
